Criação de lógicas de Torneio, sorteios e alterações de tipos

+++ app/src/main/java/com/example/p3_project/data/entities/Time.kt
@@ -5,7 +5,7 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "times")
 data class Time(
-    @PrimaryKey(autoGenerate = true) val id: Int = 0,
+    @PrimaryKey(autoGenerate = true) val id: Long = 0,
     val nome: String,
-    val torneioId: Int
+    val torneioId: Long
 )
